% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear-knitr-cache.R
\name{clear.knitr.cache}
\alias{clear.knitr.cache}
\title{Delete knitr cache}
\usage{
clear.knitr.cache(cachedir = "cache")
}
\arguments{
\item{cachedir}{The name of the directory in the present project}
}
\value{
Deletes all files in the specified directory
}
\description{
Delete everything in the project "cache" directory
Removed all files in "../cache" by default.
}
\author{
Kieran Healy
}
