% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdate.R
\name{fastDate}
\alias{fastDate}
\title{fastDate}
\usage{
fastDate(datestring)
}
\arguments{
\item{date}{A character string in the format \code{yyyy-mm-dd}}
}
\description{
A faster alternative to \code{lubridate::as_date} and \code{base::as.Date}.
}
\examples{
dates <- as.Date( seq.int( 1, to = 1e4, by = 1), origin = "1970-01-01" )
x <- as.character( dates )
fastDate( dates )
}
