% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length-zero-default}
\alias{length-zero-default}
\alias{\%|0|\%}
\title{Default value for `length(x) == 0`.}
\usage{
x \%|0|\% y
}
\arguments{
\item{x, y}{If `length(x) == 0`, will return `y`; otherwise returns
`x`.}
}
\description{
This infix function makes it easy to replace a length 0 value with a
default value. It's inspired by the way that Ruby's or operation (`||`)
works.
}
\examples{
"bacon" \%|0|\% "eggs"
NULL \%|0|\% "eggs"
}
