% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lg}
\alias{lg}
\title{length( grep( ... ) )}
\usage{
lg(x, pattern, perl = TRUE, ...)
}
\arguments{
\item{x}{a vector on which we attempt to match \code{pattern} on.}

\item{pattern}{regex pattern passed to \code{grep}.}

\item{perl}{boolean. use perl-compatible regular expressions?}

\item{...}{additional arguments passed to \code{\link{grep}}.}
}
\description{
This is a wrapper to a \code{length( grep( ... ) )}. See examples for usage.
Primarily intended for interactive, not programmatic, use.
}
\examples{
x <- c("apple", "banana", "cherry")
if( lg( x, "^ap" ) > 0 ) {
  print( "regular expression '^ap' found in 'x'" )
  }
}
\seealso{
\code{\link{re_exists}}
}
