% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lu}
\alias{lu}
\title{Number of non-NA unique elements in a vector
 
Returns the number of non-NA unique elements in a vector. A wrapper to
\code{length( unique( x[!is.na(x)], ... ) )}. 
Primarily intended for interactive, not programmatic, use.}
\usage{
lu(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{passed to \code{\link{unique}}}
}
\description{
Number of non-NA unique elements in a vector
 
Returns the number of non-NA unique elements in a vector. A wrapper to
\code{length( unique( x[!is.na(x)], ... ) )}. 
Primarily intended for interactive, not programmatic, use.
}
