% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{midrange}
\alias{midrange}
\title{Midrange}
\usage{
midrange(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric scalar vector.
}
\description{
Point halfway between the min and the max.
}
\details{
The point halfway between the min and the max. The mid-range minimizes the maximum distance to a set of points.
}
\examples{
midrange(iris[,1])
midrange(iris[,2])
midrange(iris[,3])
midrange(iris[,4])
}
