% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.r
\name{my.colors}
\alias{my.colors}
\title{Color-blind friendly palette}
\usage{
my.colors(palette = "cb")
}
\arguments{
\item{palette}{Choose "cb", "rcb", or "bly". cb is the Winston
Chang color blind palette; rcb is that palette in reverse; bly
puts the yellow and blue in the palette first.}
}
\value{
Variations on an eight-color color-blind friendly palette.
}
\description{
Color-blind friendly palette
From \url{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}
}
\author{
Kieran Healy
}
