% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{\%nin\%}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector of items}

\item{y}{vector of all values}
}
\value{
logical vecotor of items in x not in y
}
\description{
Inverse Value Matching
}
\details{
Complement of \code{\%in\%}. Returns the elements of \code{x} that are not in \code{y}.
}
\author{
Kieran Healy
}
