% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Rep Mat Function}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{a numeric matrix}

\item{m}{number of row-wise replications}

\item{n}{number of column-wise replications}
}
\value{
numeric matrix with X replicated n-times across and m-times down
}
\description{
This function takes a matrix X and replicates it n times column-wise and m times row-wise, similarly to repmat in Matlab
}
\examples{
set.seed(1)
x = matrix(rnorm(6),2,3)
x
repmat(x,3,2)
}
\keyword{repmat}
