% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shell_transformer}
\alias{shell_transformer}
\title{System convenience functions ------------------------------------------------
Glue to shell functions}
\usage{
shell_transformer(code, envir)
}
\description{
System convenience functions ------------------------------------------------
Glue to shell functions
}
