% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_github.r
\name{update_github}
\alias{update_github}
\title{update_github}
\usage{
update_github()
}
\value{
none
}
\description{
Update installed github packages
}
\details{
Update installed github packages
}
\author{
https://stackoverflow.com/questions/32538052/update-all-packages-from-github
}
