% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize Function}
\usage{
winsorize(x, fraction)
}
\arguments{
\item{x}{numeric vector of real numbers}
}
\value{
numeric vector of x with all elements below the franctionth percentile truncated to the limiting value from the top and bottom
}
\description{
This function truncates vector to the fractionth-ile elements from the top and the bottom
}
\examples{
set.seed(1)
x <- rnorm(10)
x <- sort(x)
x
winsorize(x,0.2)
}
\keyword{winsorize}
